Network.Google.AppsCalendar.Converter.ICalFormat

Description

This module provides functions to map subparts of a VEvent to Text. Especially the part of a Recur object is useful here for the conversion of a VEvent.

The code is basically a copy of the Printer module, which does not export functionality to print the subparts of a Text.VEvent object.

Typeclass to map subparts.

class IsValue a where

A typeclass that specifies that the given type can be printed as specified by the iCal format.

Methods

printValue

Arguments

:: a
The given element to print.

-> ContentPrinter ()
a builder that will print the specified object.

Convert the given item to a ContentPrinter that prints it as specified by the iCal format.

Types to make printing more convenient.

type ContentPrinter = RWS EncodingFunctions Builder Int

A type synonym of a way to print data. This makes use of the Builder to generate the stream and the EncodingFunctions as state.

data EncodingFunctions

A datatype that is used as state to construct a textual representation of the objects.

Printing properties.

class IsProperty a where

A typeclass that is used to print items as properties of a parent element.

Minimal complete definition

printProperty

Methods

printProperty

Arguments

:: a
The given element to print.

-> ContentPrinter ()
A builder that will print the given property. | Convert the given element to a Text object that is the iCal format of that element.

print the element as specified by the iCal format.

propertyToText

Arguments

:: a
The given element to convert to Text.

-> Text
The textual presentation of the element as a property.